Add infrastructure for count interface support
Add a count-variant helper library for the mpi/c/ library, to
begin adding support for MPI_Count count variants.  We assume
that there will genearally be high code reuse between the int and
MPI_Count variants and that, like the profiling setup, we will
want to use the same source file for both variants (although this
is not strictly required).

Implement MPI_Send_c as an example.