DRAFT: build: Add static component pointers
With GCC's link time optimization, the compiler ends up seeing two
declarations of MCA component structures.  One in the component
itself, which has potentially a component-specific type, and one
in static-components.h, which is just declared as a base MCA
component struct.  The compiler doesn't like this.

This patch adds a level of indirection used by the static component
lookup code.  static-components.h now declares the existance of a
mca_component_t * and the list is of pointers to pointers.  The
MCA base framework has to do an extra dereference, but this is only
in the initialization path.  We also then have to have the actual
pointer exist, so add a macro MCA_BASE_COMPONENT_INIT that we need
to add to every component definition to create the pointer (and
perhaps do more initialization later).